Add alert bell GIF (#452)
* Add alert bell GIF

* Extends frame display time for GIF viewer

Allows control over how long each frame is displayed.

Updates the alert bell to show for a longer duration by introducing a frame multiplier.

Adds validation to ensure the multiplier is valid.

* Multiplies canvas count by frame multiplier

Multiplies the number of available canvases,
which resolves an issue where the alert bell
animation was too slow.

* Handles invalid frame multiplier values

Ensures that the frame multiplier is not less than 1,
setting it to a minimum value if needed.

Logs the determined canvas count for better debugging
and visibility.

* Reduces alert bell GIF frame multiplier

Decreases the frame multiplier for the alert bell GIF.

The previous value caused the animation to play for an
excessively long duration, which was not ideal for quick notifications.

The new value ensures a more responsive and noticeable alert.

* Adds commit selection for rpi-rgb-led-matrix

Allows specifying a commit hash to use for the
rpi-rgb-led-matrix submodule, instead of always
checking out the master branch. This provides
more flexibility and control over the version
of the library used.

* Resets the RGB matrix repo for clean builds

Ensures a clean build by removing the existing
rpi-rgb-led-matrix directory before cloning.

This prevents potential issues caused by lingering files
or incorrect states from previous builds, ensuring a
consistent and predictable environment.

* Fixes GIF frame count

Corrects an issue where the GIF viewer iterates based on the incorrect number of frames.

It now uses the image's actual frame count for iteration, ensuring all frames are displayed.

* Grants process permission for priority control

Allows the service to adjust process priority.

The service now has the capability to set its own
scheduling priority, which can be useful when timing
is crucial for triggering alerts.

* Handles image transparency

Ensures transparency in images is correctly handled by compositing images with transparency onto a black background before displaying them on the LED matrix.

This prevents issues with displaying images that have transparent areas.

* Removes and re-adds alert bell animation

Replaces alert bell animation.
This likely updates or corrects the animation.

* Removes duplicate asset

Removes a duplicate animated asset from the repository.

* Adjusts GIF frame multiplier for better display

Tweaks the frame multiplier for the alert bell and door GIF viewers.

Increases the multiplier for the alert bell to ensure it's more visible and appropriately timed and reduces the multiplier for another to ensure the animation isn't too fast.

This provides a better visual experience.

---------

Co-authored-by: Will Garside <worgarside@gmail.com>